reestructured search component to avoid code repetition
